dom/tests/mochitest/notification/test_bug931307.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Bug 931307</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<pre id="test">
<script type="application/javascript"><!--

SimpleTest.waitForExplicitFinish();
new Notification("");
var promise = Notification.get();
promise.then(
  function onSuccess() {
    ok(true, "No crash!");
    SimpleTest.finish();
  },
  function onFailure() {
    ok(false, "Should not get an error in promise callback");
  }
);

</script>
</pre>
</body>
</html>
